feat: finally file serving!!, dir serving soon™

+++ justfile
@@ -10,14 +10,18 @@ reset := "\\x1b[0m"
 include_dir := "include"
 src_dir := "src"
 out_dir := "out"
+asm_dir := "asm"
 bin_dir := "bin"
 include_flags := '-I' + include_dir
+link_flags := ' -lmagic'
 debug_shared_flags := '-ggdb -g -Og -fsanitize=address,undefined,leak -fno-sanitize-recover=all -fno-omit-frame-pointer -fno-optimize-sibling-calls -fsanitize-address-use-after-scope -fno-common -std=gnu11'
 debug_compile_flags := debug_shared_flags + ' -Wall -Wextra -Wpedantic -Wno-unused-parameter'
-debug_link_flags := debug_shared_flags + ' -static-libasan'
-release_shared_flags := '-Ofast -std=gnu11'
+debug_link_flags := debug_shared_flags + link_flags + ' -static-libasan'
+release_shared_flags := '-O2 -std=gnu11'
 release_compile_flags := release_shared_flags
-release_link_flags := release_shared_flags
+release_link_flags := release_shared_flags + link_flags
+debug_static_link_flags := debug_link_flags + ' -static'
+release_static_link_flags := release_link_flags + ' -static'
 
 # Default justfile target (lists all available targets)
 default:
@@ -52,6 +56,7 @@ compile type="debug" force="false" threads=num_cpus():
       exit 0
     fi
 
+
     echo -e "Using {{ red }}{{ threads }}{{ reset }} threads"
     echo -e "Target: {{ green }}{{ type }}{{ reset }}"
     if [[ {{ type }} == "debug" ]]; then
@@ -63,7 +68,7 @@ compile type="debug" force="false" threads=num_cpus():
     fi
 
 # Links all the object files in the out directory based on the type of build (debug or release)
-link type="debug":
+link type="debug" static="dynamic":
     #!/usr/bin/env bash
 
     [[ -d {{ out_dir }} ]] || just compile {{ type }}
@@ -78,8 +83,14 @@ link type="debug":
       done
     else
       for file in {{ out_dir }}/*-release.o; do
-        if [[ $file -nt {{ bin_dir }}/httiny ]]; then
-          WILL_LINK=true
+        if [[ {{ static }} == "true" || {{ static }} == "static" ]]; then
+          if [[ $file -nt {{ bin_dir }}/httiny-static ]]; then
+            WILL_LINK=true
+          fi
+        else
+          if [[ $file -nt {{ bin_dir }}/httiny || $file -nt {{ bin_dir }}/httiny-static ]]; then
+            WILL_LINK=true
+          fi
         fi
       done
     fi
@@ -91,23 +102,34 @@ link type="debug":
 
     FILES=$(find {{ out_dir }} -name "*-{{ type }}.o")
 
-    echo -e "Linking {{ green }}{{ type }}{{ reset }}"
-    echo -e "Files to link: {{ green }}"
-    printf "%s\n" "${FILES[@]}" | sed 's/^/    /'
-    echo -e "{{ reset }}"
+    if [[ {{ static }} == "dynamic" || {{ static }} == "false" ]]; then
+      echo -e "Linking {{ green }}{{ type }}{{ reset }} dynamically"
+      echo -e "Files to link: {{ green }}"
+      printf "%s\n" "${FILES[@]}" | sed 's/^/    /'
+      echo -e "{{ reset }}"
 
-    if [[ {{ type }} == "debug" ]]; then
-      gcc {{ out_dir }}/*-debug.o {{ debug_link_flags }} -o {{ bin_dir }}/httiny-debug -fuse-ld=mold
+      if [[ {{ type }} == "debug" ]]; then
+        gcc {{ out_dir }}/*-debug.o {{ debug_link_flags }} -o {{ bin_dir }}/httiny-debug -fuse-ld=mold
+      else
+        gcc {{ out_dir }}/*-release.o {{ release_link_flags }} -o {{ bin_dir }}/httiny -fuse-ld=mold
+      fi
+      echo -e "Linking {{ green }}{{ type }}{{ reset }} dynamically complete"
     else
-      gcc {{ out_dir }}/*-release.o {{ release_link_flags }} -o {{ bin_dir }}/httiny -fuse-ld=mold
+      echo -e "Linking {{ green }}{{ type }}{{ reset }} statically"
+      echo -e "Files to link: {{ green }}"
+      printf "%s\n" "${FILES[@]}" | sed 's/^/    /'
+      echo -e "{{ reset }}"
+      gcc {{ out_dir }}/*-release.o {{ release_static_link_flags }} -o {{ bin_dir }}/httiny-static -fuse-ld=mold
+      echo -e "Linking {{ green }}{{ type }}{{ reset }} statically complete"
     fi
 
-    echo -e "Linking {{ green }}{{ type }}{{ reset }} complete"
-
 # Builds the project with the given type and force flags
-build type="debug" force="false" threads=num_cpus():
+build type="debug" force="false" threads=num_cpus() static="dynamic":
     just compile {{ type }} {{ force }} {{ threads }}
-    just link {{ type }}
+    just link {{ type }} {{ static }}
+
+build_static force="false" threads=num_cpus():
+    just build release {{ force }} {{ threads }} static
 
 # Alias for building the project in debug mode
 debug force="false" threads=num_cpus():
@@ -116,6 +138,9 @@ debug force="false" threads=num_cpus():
 # Alias for building the project in release mode
 release force="false" threads=num_cpus():
     just build release {{ force }} {{ threads }}
+
+release_static force="false" threads=num_cpus():
+    just build_static {{ force }} {{ threads }}
 
 # Runs the project with the given type and force flags
 run type="debug" force="false" threads=num_cpus():
@@ -128,9 +153,44 @@ run type="debug" force="false" threads=num_cpus():
       ./bin/httiny
     fi
 
+generate_asm force="false" threads=num_cpus():
+    #!/usr/bin/env bash
+    shopt -s globstar
+
+    [[ -d {{ asm_dir }} ]] || mkdir -p {{ asm_dir }}
+
+    WILL_COMPILE=false
+
+    for file in {{ src_dir }}/**/*.c; do
+      if [[ $file -nt  {{ asm_dir }}/$(basename "${file%.c}").S ]]; then
+        WILL_COMPILE=true
+      fi
+
+      if ! [[ -f {{ out_dir }}/$(basename "${file%.c}").S ]]; then
+        WILL_COMPILE=true
+      fi
+    done
+
+    if [[ {{ force }} == "force" || {{ force }} == "true" ]]; then
+      WILL_COMPILE=true
+    fi
+
+    if [[ $WILL_COMPILE == false ]]; then
+      echo -e "Compile: Nothing to do"
+      exit 0
+    fi
+
+
+    echo -e "Using {{ red }}{{ threads }}{{ reset }} threads"
+    echo -e "Target: {{ green }}asm{{ reset }}"
+    find {{ src_dir }} -name "*.c" -print0 | xargs -0 -P{{ threads }} -n1 \
+      sh -c 'if [[ "$1" -nt "{{ out_dir }}/$(basename "${1%.c}").S" || {{ force }} == "true" || {{ force }} == "force" ]]; then echo -e "Compiling {{ green }}$1{{ reset }}..."; gcc {{ include_flags }} {{ release_compile_flags }} -fverbose-asm -S "$1" -o "{{ asm_dir }}/$(basename "${1%.c}").S"; fi' sh
+
 # Cleans the out directory
 clean:
-    rm -rf {{ out_dir }}
+    [[ ! -d {{ out_dir }} ]] || rm -rf {{ out_dir }}
+    [[ ! -d {{ bin_dir }} ]] || rm -rf {{ bin_dir }}
+    [[ ! -d {{ asm_dir }} ]] || rm -rf {{ asm_dir }}
 
 # Uses bear to generate compile_commands.json
 bear:
